requirements.txt
django_heroku==0.3.1
gunicorn==20.0.4
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.3
pandocfilters==1.4.2
pickleshare==0.7.5
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
requests==2.23.0
requests-toolbelt==0.9.1
scikit-learn==0.22.2.post1
scipy==1.4.1
fastai==1.0.60
